+++ src/main/java/com/soolsul/soolsulserver/bar/businees/BarQueryService.java
@@ -1,8 +1,8 @@
 package com.soolsul.soolsulserver.bar.businees;
 
-import com.soolsul.soolsulserver.bar.dto.response.BarLookupResponse;
-import com.soolsul.soolsulserver.bar.dto.request.BarLookupServiceConditionRequest;
-import com.soolsul.soolsulserver.bar.dto.response.BarsLookupResponse;
+import com.soolsul.soolsulserver.bar.businees.dto.BarLookupResponse;
+import com.soolsul.soolsulserver.bar.businees.dto.BarLookupServiceConditionRequest;
+import com.soolsul.soolsulserver.bar.presentation.dto.BarsLookupResponse;
 import com.soolsul.soolsulserver.bar.persistence.BarQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
