+++ src/main/java/com/soolsul/soolsulserver/bar/businees/dto/BarLookupResponse.java
@@ -1,4 +1,4 @@
-package com.soolsul.soolsulserver.bar.dto.response;
+package com.soolsul.soolsulserver.bar.businees.dto;
 
 import java.time.LocalDateTime;
