+++ src/main/java/com/soolsul/soolsulserver/bar/businees/dto/BarLookupServiceConditionRequest.java
@@ -1,13 +1,14 @@
-package com.soolsul.soolsulserver.bar.dto.request;
+package com.soolsul.soolsulserver.bar.businees.dto;
 
 import javax.validation.constraints.NotEmpty;
+import java.util.List;
 
-public record BarLookupConditionRequest(
+public record BarLookupServiceConditionRequest(
         @NotEmpty double southWestLongitude,
         @NotEmpty double southWestLatitude,
         @NotEmpty double northEastLongitude,
         @NotEmpty double northEastLatitude,
-        @NotEmpty String barMoodTagNames,
-        @NotEmpty String barAlcoholTagNames
+        @NotEmpty List<String> barMoodTagIds,
+        @NotEmpty List<String> barAlcoholTagIds
 ) {
 }
