--- src/main/java/com/soolsul/soolsulserver/bar/dto/response/BarsLookupResponse.java
@@ -1,8 +1,0 @@
-package com.soolsul.soolsulserver.bar.dto.response;
-
-import java.util.List;
-
-public record BarsLookupResponse (
-        List<BarLookupResponse> barLookupResponses
-) {
-}
