+++ src/main/java/com/soolsul/soolsulserver/bar/facade/BarQueryFacade.java
@@ -3,9 +3,9 @@ package com.soolsul.soolsulserver.bar.facade;
 import com.soolsul.soolsulserver.bar.businees.BarAlcoholTagService;
 import com.soolsul.soolsulserver.bar.businees.BarMoodTagService;
 import com.soolsul.soolsulserver.bar.businees.BarQueryService;
-import com.soolsul.soolsulserver.bar.dto.request.BarLookupConditionRequest;
-import com.soolsul.soolsulserver.bar.dto.request.BarLookupServiceConditionRequest;
-import com.soolsul.soolsulserver.bar.dto.response.BarsLookupResponse;
+import com.soolsul.soolsulserver.bar.presentation.dto.BarLookupConditionRequest;
+import com.soolsul.soolsulserver.bar.businees.dto.BarLookupServiceConditionRequest;
+import com.soolsul.soolsulserver.bar.presentation.dto.BarsLookupResponse;
 import com.soolsul.soolsulserver.menu.alcohol.service.AlcoholCategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
